The recent article on how the world savers at OpenAI make tilts me a little. Going to good schools and getting good grades sure worked for them.
I'll have to make do with actual skills rather than hype.

I've really narrowed the path to tier one with 3 key ideas regarding how get rid of state aliasing.

1) Explicit modulation - train a net on top of a net to predict novelty and use that to modulate learning.

2) Implicit modulation through memory - while the states might alias the hidden state might be used to differentiate them. The way to stabilize will be through adding memory and depth to RNNs.

3) Deep implicit modulation - RL algorithms are fundamentally broken and the best way of doing RL will be to let the net do everything. Hence the way to train value function would be recursive by deep propagation through the world model and the controller.

The third one would actually be needed for reasoning, so it will have to be faced at some point, but after thinking through it for a while I am starting to lean towards the second one.

I am overthinking the whole gradient propagation thing. It is not really as much of a problem as I make it out.

The real issue is making sure that there is significant differentiation in states which will allow the training to go smoothly.

Quantile regression will take care of the large differences in reward inputs.

It is useful to go into extremes for the sake of generating ideas.

I sort of feel like the other two points are a bit extreme. It is not like they are at all invalid, but I personally I'd want to exhaust the architecture path first.

I needed this. I needed to go through this tumult in order to form my own opinion. I have a rather large variety of hypotheses lodgen in my mind now which will allow me to place future breakthroughs in context.

I have enough inspiration gathered to be confident in my own path and my own approach.

The second option which is enchancing the memory capabilities either through depth or other mechanisms like differentiable plasticity will without a doubt work.

I am sure of it.

It will get me to tier one and will let me continue climbing for there. I just need to go through it first.

It is time to continue the buildup.

Getting to tier one will be the most strenous part of this whole thing without a doubt. In fact the whole creation of Spiral and what came before can be considered a part of it.

After I manage to reach the next level, I will be able to to slow down my pace a little I think as at that point I'll have the computer taking up the slack.

Well, that aside, I am thinking of how I am going to do the encoder.

I do not think I have much choice here - in order to do union types properly, I am going to have to extend the language...

But that would solve just one of the issues that I have in mind.

...Actually, this problem is really hard.

It is going to require some deep thought.