Today I was just too distracted by papers and videos. Let me at least skect out the solution today and tomorrow I will extend the language.
I am enjoying this. It is like being on the hunt.

This plastic NN stuff is going to be decisive down the road and Spiral is uniquely positioned to experiment with it. There is finally a reason to ditch the BLAS library.

But for now let me make the encoder.