The sum type works, but I am going to have to try nested types before I can make sure.
I am still thinking about whether integers should be passed as one hot vectors. I think there should be a better choice.

Well, I'll think about it when it comes to that.

I need to keep in mind that I do not need to make this stuff perfect for every use case. Originally I only really needed this serializer function because of the union types and this is what it should be doing.

I'll do other kinds of schemes when the need for them arises.

Let me do one more test and then I can actually move to making the test.