I have the encoder dealt with.
Brilliant. This will really help to put a lid on errors when passing union types to the GPU.

I was really uncomfortable with making serialization functions by hand back when I was doing it in F#.

I kind of want to take a little break here to read some papers on differentiable structures by Thomas Miconi. I also want to read the recent paper that was posted on dissecting the Nvidia architecture.

I need to let this lie down anyway before I move to the next steps which would be making the input layer for RL.

After that is done I'll have to make the decoder somehow. Let me put some thought towards figuring out exactly how I want to do this.