Let me start work on rl_output.
I am still getting my bearings. I feel out of touch with the library after so long, but basically getting the inputs and output done right is all I need to implement the first version of Q learning.

It won't be long until the net is crushing the random player. I'll be done in a few days at most.