Time to do some programming. I got up an hour ago and have been reading manga all this time. Time to finish the decoder and then I'll maybe watch some videos from the Simons Institute.
In truth, there is not that much work left.

Originally before differentiable plasticity came to the scene I just intended to implement dense and dilated connections and that would have been the perfect architecture. Dense connections for the vertical dimension and dilated for the temporal. There is not exceeding that.

Now that DP is here I have more work to do, but not that much. It is plausible that two months will be enough to get through all of it.

So this is most likely a contributing factor in why my sleep has gone a lot better for the past few weeks. Yeah, there is some drudgery, but I'll get through it with some effort.

There are still issues with Q learning that need to be solved, but while I do research and wait for something new to come out to push me all the way to tier one, I'll finish the build up.

Unlike Deepmind, in order to beat online poker, I need to be a lot better than the usual setup. I really need the algorithm to be a lot better than it would be had I invited other players to my specially prepared kill zone. Here I will be going into their home turf.

...Actually, let me take a break now instead of later. I'll get some exercise and warm myself up before starting programming.