https://stackoverflow.com/questions/14093601/effect-of-using-page-able-memory-for-asynchronous-memory-copy/
So to make this happen all I need to do is chance the standard memcopy to an async one. No problem. I am not going to use async memory beucase it would be a pain in the ass though. What I really wanted most was to not have it block host.

Let me take a look at the CudaTensor module again.