I sketched out the RL functions here. I wrote them two hours ago before I left for breakfast and ready to resume work now.
The way I wrote it out, I got a really good feeling from them.

Yeah, the code really feels good now.

There has been a slight chance in policy in the library. The place to pass in the reward is definitely in the backwards function.

This is a departure from how it is done in supervised learning. Without a doubt, the right place to do it..

...Er, is it? What if I have multiple actions?

Let me think about this a little.