Let me commit this before I get distracted any more. Whenever I finish a chunk I fall into the tendency to start surfing the Internet at random.
At any rate, I think the recent additions make sense. If I need multiple actions, then they will have to be done recurrently.

...I think this about covers all the design aspects that needed to be considered. I think now all the pieces are in place for me to really start work on this.

The decoder won't be too hard, but the action selectors will be a bit tricky. But since I already went through the sampler, I'll be able to use it as a template.

...Well, let me just aim to do the encoder for today.