As it turns out there were bugs in foldl2 and foldl3. But at any rate, this should do for foldl_map series of functions.
This was much easier than I thought it would be. Well, unless there is some kind of error, but I'll leave that for when it comes to it. I can't test every little thing.

Right now I am thinking how to the last major step. That would be be the action selector. I'll start of with the square error one for starters.

Since there is still time in the day, I'll some more effort into doing them.