I will call it a day here. I really can't go any further that this. I'll test the selector function and everything else tomorrow.
If everything goes fine, I'll manage to implement a NN based for the first time in two days. I do not think I'll manage it tomorrow, but the exact moment where I do this is not so far away.

Once I break the ice, it really won't take me that long to get up to full power. Putting in the replay buffer, distributional RL and recurrent nets is around a week of work in my estimate. At most two.

Two is more realistic in real time since I won't be programming in its entirety, but will want to play around for a bit.

Despite all my distraction it feels like I managed to do quite a bit today.

Good work, me. Now it is time for my well deserved rest. As usual I pass the baton to the me of the future.