https://igi-web.tugraz.at/PDF/241.pdf Deep rewiring: training very sparse deep networks
https://arxiv.org/pdf/1712.06560.pdf
Improving Exploration in Evolution Strategies for Deep Reinforcement Learning via a Population of Novelty-Seeking Agents

Let me get breakfast. These are the last two papers. After that I will gather my will and simply decide that MC updates are right. This has been bothering me for a long time. I mean regarding how to stably propagate gradients as they are.

Really, the only answer is recurrence and rich internal states. This can be the only answer for why standard Q learning is unstable.

There are other aspects of RL such as training on top of a generative model, but there simply isn't any way around the core idea of doing Monte Carlo updates in order to do prediction.

It has to be right - it simply has to. I've been trying to reason my way around a problem that does not exist. There aren't any alternatives to this. At least not for the context I am trying to solve.

For poker, the basic plan of doing MC updates combined with architectural innovations is the one I should pursue. I want model-free learning here. Levine's analogy of how pitchers catch balls has struck a chord with me and I find myself agreeing with it.

For very long time horizons and model based learning there will probably be alternatives in the future, but forget that for now. I'll deal with it when the need arises. For model-free learning, there are only so many options at my disposal.