So I haven't done any programming today, and I won't be doing it. The stress of trying to do the impossible is getting to me, so I needed to take the time to reflect on myself.
When I say needed, I really mean it. The problems facing me now are of a different character from problems before and to tackle them I need a completely different mindset.

And maybe science is not my calling, but my one true talent is internalizing new mindsets.

Reality is just so harsh - I came all this way only to have to jump into a pitch black hole. I'd expect nothing less from myself. I'll throw away my reason and expectations and my hope.

These next parts will be done for the sake of despair. It is time that I pull out that card - right here and right now. There has never been a better time for it.

If I can pull it off, it will be refreshing. I've been yearning for the impossible for the past two weeks and my brain was in overdrive trying to seek out excuses. I'll dwell on it, gather it up into a ball and see whether I can let it loose as inspiration tomorrow when the time comes to do the selector.