Ugh, I've been making post left and right ever since I got up. It feels like being on a roller coaster. I also got mail that I am AI grant finalist and now I have ot make a video.
Well, I'll find that webcam later.

I got reply by the Dora author saying they did not remove the biases. That is very interesting as it means the network will not unlearn the optimistic bias quickly. I had not expected that would be the case.

They haven't done any comparisons with Rmax and the rest was roughly what I expected. They ran the Freeway test 5 times, the base of the log does not matter, the performance comment was meant on a per episode basis.

If I had to guess, I'd bet that having a stream just for adding an exploration bonus is the right way of doing this (rather than adding a large positive constant and removing the bias terms.)

Even though it does not really make sense, having critics, double Q streams, distributional rewards and such really does improve performance, so having a stream for optimization will be better than cramming the optimism into Q values.

So I will go with Dora as my exploration method.

Now...the only issue that remains is cracking the first tier.

https://www.reddit.com/r/MachineLearning/comments/8f170c/d_would_metalearning_work_for_training_generative

I think at this point, the time is here to doubt some very basic things like whether the inputs are being inserted in the right order.

Ultimately, all the information is in the inputs, but they are secondary to the optimization scheme. It seems obvious that if the places were swapped, the network would be receiving much more information.

Assuming that works and drives a wedge through the door called generalization, then the next step would be to make the weights symmetric and that would suddenly make neural nets much more biologically plausible. At the very least, there is a lot reason to doubt that rewards should come in the outputs in RL settings. The amount of information they carry is basically nil.

Now, let me leave that aside and do some studying of Bayesian RL and the rest of such things.