The more I study, the closer I get to the true principles of intelligence.
I've pretty much decided. Metatraining for invertibility is what is needed to crack tier one. Without a doubt it is the missing piece.

https://arxiv.org/abs/1705.08664
Towards Understanding the Invertibility of Convolutional Neural Networks

https://openreview.net/forum?id=HJsjkMb0Z
i-RevNet: Deep Invertible Networks

It does not feel like this is quite it, but I will take a look.

I just need to go a little bit further and pieces will fall into place. Time to resume.