I'll do the required 2m video for the AIgrant challenge soon.
I feel a bit nervous. Let me just say that damn the whole field to hell. The parts of DL that actually make sense ot me now are those that do not even exist.

Just now I've realized that the dynamics models do not make sense. It is for the simple reason that it does not necessarily make sense to reproduce the inputs directly.

Levine's group did a very interesting thing where they passed an image into the hidden layers and then used that hidden representation as the target for a robot to reach.

That is absolutely brilliant and explains how sexual attraction could work in humans.

But if we are predicting dynamics, when humans do reasoning they are almost certainly not predicting the raw things, but they are going through high level pathways instead.

I've read some recent papers on model based RL and yes, they have much better sample efficiency and work to an extent.

But I cannot sense the force of principle behind them.

The more I think about algorithms like MAML and Reptile the more convinced I become that the whole field has it all backwards and that labels and inputs should have their places swapped.

The inputs should come at the top where they would be the most informative and the labels should come at the bottom to condition the net.

That is what would make the most sense for the sake of maximizing information.

I mean the labels themselves simply have nothing.

So this is the first absolute principle that makes sense.

I am watching videos, reading papers and learning new things, but the current methods are just hacks.

I've been so foolish. Really until the 10th of this month I really did not care about this at all.

Now that I am making a concious decision what to use, being able to imagine the correct form almost makes it physically painful to do this.

How could I really for something that is 1000x weaker than I'd prefer it to be. This cannot stand.

I'll do the basic thing. I'll use Dora for exploration and MC updates. I'll add plastic neurons and Reptile and quantille regression in addition to that.

But that will most like be the end of it.

I'll get something going with that and then I will just wait and seek out the key to the next tier myself if it does not seem to be forthcoming.

Let me do the video.