Actually, the TDM paper is rather good. I need to reflect more on it.
The way they set rewards and the way they define the Q function is some interesting engineering. It seem that they also do lookahead in order to make sure the action is optimal.

Actually, that lookahead is something I should definitely consider for poker. Since the individual decisions just matter so much maybe training a recurrent model over the hidden states given the action would not be a bad idea.

https://www.reddit.com/r/reinforcementlearning/comments/8f5obs/temporal_difference_models_modelfree_deep_rl_for/

I'll post a review on this thread a bit later.