Honestly, I feel like my journey really resembles that off the backprop algorithm as I am pushed from place to place. There are all sorts of advancements to be sure, but all these huge gains in efficiency are really because of the lack of a real unsupervised learning.
Bacause RL is really like sipping through a mini-straw, every single gain in information has an absolutely outsized effect on the performance of the whole system.

...

Ah, forget it. There is no point in getting obsessed about it.

Ultimately my strategy should be to be to maximize the advantages that exist and do what I can with the resources on hand. I cannot possibly see through the effects of optimization programs ahead of time.

Maximize the signal. Use self play.

Poker is ideal for that sort of thing.

Let me see if I can do a bit more before that day is over.