With this I am done for the day. The player is essentially halfway done even though I am just getting started with the whole thing.
Being able to serialize states and actions into one hot format and back is a huge thing.

If I hadn't done it like that that particular part would have been a huge hotbed of errors across a range of different games. It would have been an achor I'd have to drag all the way through the codebase.

This is the real power of refection.

Now...I made a mistake in the design.

The backwards pass was not meant to receive inputs after all in the form of reward.

The way it is done now will make it really hard to understand how the optimization process is going.

....Actually, that was just a thought. I need to think more how I want to organize this whole thing.

Tomorrow I will make the first steps towards completing the play and plugging it into the simulator.

I did well today, so it is time to rest.