I read a bunch of papers, but they are mostly filler. They haven't introduced any new principles for me. The Noisy Nets in particular made me realize that Freeway really is not a challenging exploration environment and that Dora really needs more experiments.
The TDM paper yesterday was trully interesting, but picking random goals is certainly not the ideal. I just still cannot bring it together.

This is hard, really hard. The reason why RL algorithms have such insane performance differences is because they are crazily starved so any improvement gives exponential benefits to performance.

I worry, and worry and think, but it is not going to get me anywhere.

Ironically, I am missing the one thing that would make my algorithms work - a principled way of doing exploration which leaves me uncomfortably reliant on the field to make the vast majority of real progress.

Such despair. This world is certainly not the place to relax in.

Very well. I did my reading for the day. Let me put in some work, rather than worry about things outside my control.

What should be my goal for today?

...I have some issues with passing in rewards.

I think that maybe just for the action layer, I should avoid term casting it. It is just too annoying.

Or maybe....

You know this is not like me. It just occured to me that if I was trully serious about this, instead of going around in circles and wilting from the challenge in front of me, that I'd just accept my own weakness and take up the challenge.

I already agree that the brain is doing metalearning. And metalearning is what I myself is what I should be doing. Exploration is metalearning. Generalization is metalearning.

Is implementing Reptile and differentiable plasticity really the end of my ambition? Do I really just want to finish this and relax?

Is programming really just this? Is this what Spiral's power amounts to?

I already know that there is a far higher level than this. I've spent like 1.5 years working on a language.

Why don't I once more repeat the Spiral quest?

Instead of aiming for just speed, I should make a much, much better library that can accomodate metalearning needs.

The reality of the present situation is that metalearning has not been explored at all.

I have absolutely no idea how I am going to do it, but some things are clear.

1) I should just stop dawdling with poker and finish the damn players and implement all that is needed according to the original plan.

I can't keep delaying this forever.

2) Instead of staring at the pitch black abyss, grand majestic goals are what fire me up. Programming itself is about generalization. Feeling that I've reached my limits is the main thing holding me back.

During the night I've been thinking and the one thing that is sure is that only a very tiny fraction of the neural nets true power is being tapped into right now.

What we have now is not real.

https://www.youtube.com/watch?v=LnXgs73OUjE&index=1&list=PLgKuh-lKre115020ug_gdmRroIfiQsPL4&t=1s
The Prefrontal Cortex as a Meta-Reinforcement Learning System

As I am still in the process of forming my principles, let me watch this talk. I already watched a similar one given by the author a year ago, but I might as well do a bit more studying.

Despair pushes away from the present and hope draws one towards the future. That is the true principle.

I won't be able to move forward unless both are well calibrated.

I have so many things holding me back. With a much greater level of programming skill I could just break through this. The reason why I am having so much trouble making progress is because has started to feel just like janitorial work.

I cannot accept this.

I will focus and gather than though and let it loose once more.

Let me watch the talk.