"8:15am. I am up. I feel like reading about intrinsic curiosity rather than jumping into programming right away.
https://arxiv.org/abs/1710.03641
Continuous Adaptation via Meta-Learning in Nonstationary and Competitive Environments

Let me also read this.

> In this paper, we develop a gradient-based meta-learning algorithm similar to (Finn et al., 2017b) and suitable for continuous adaptation of RL agents in nonstationary environments. More concretely, our agents meta-learn to anticipate the changes in the environment and update their policies accordingly.

Oh, quite interesting.

> We evaluate our meta-learning agents along with a number of baselines on a (single-agent) locomotion task with handcrafted nonstationarity and on iterated adaptation games in RoboSumo. Our results demonstrate that meta-learned strategies clearly dominate other adaptation methods in the few-shot regime in both single- and multi-agent settings.

Oh, it is Monday. I'll read Tog after I am done with this paper.

8:55am.
> In RoboSumo, rewards are naturally sparse: the winner gets +2000, the loser is penalized for -2000, and in case of a draw both opponents receive -1000 points.

This is interesting. A draw is penalized.

9:05am. This paper is interesting. Ah, fuck it, I will study as long as I feel like. And as expected, they covered everything here. RNNs + meta updates is the best strategy.

9:15am.

> Additionally, our meta-updates were based on the gradients of a surrogate loss function. While such updates explicitly optimized the loss, they required computing second order derivatives at training time, slowing down the training process by an order of magnitude compared to baselines.

An order of magnitude? Yikes.

I'd expect a slowdown even for Reptile as it needs to duplicate the weights.

The LSTM + meta updates approach is absolutely dominant.

9:25am. This makes it really easy for me decide whether I want to use RNNs + metalearning or not. I was already intending to, but this is a concrete piece of evidence in the superiority of the approach in a time where data is sparse.

Let me read Tog and then I'll look up some stuff on intrinsic motivation. It is not the same as exploration. Right now I am trying to come to some conclusion regarding what to do about goal formation. For poker in particular adversarial dynamic should be enough to drive development, but I won't be satisfied with just this."