https://pathak22.github.io/noreward-rl/resources/icml17.pdf Curiosity-driven Exploration by Self-supervised Prediction
This paper really feels a lot more clear to me than almost two months ago. I really wish they did not train the features as a part of the inverse model. I want to follow up on the related work. `Loss is its own Reward` is something I should revisit. I should also mail the authors to see if they tried learning the features as a part of the control or in a unsupervised fashion.

This is something that did not occur to me to wonder about last time, and it might seem sensible in the context, but in poker for example, the action is perfectly predictable from the two states, so it makes no sense that the inverse model should be used to train the visual system.

...Actually it should be perfectly predictable in Doom and Mario as well. What the hell?

Er, I need to spend some time thinking about this.