I am done with the poker5 test. I got the DQN player to work.
A few notes:

- Linear combinations of features works suprisingly well.
- Layer norm just makes everything worse.
- A net with a single hidden layer works well, but not any better than linear.
- The whole thing is 50-100x slower per iteration than the tabular case which is not surprising. And it is 100-300x slower in the hidden layer case.
- Biasing the rewards does not seem to be particularly helpful. The positive and negative biases seems to unlearned quickly.
- This whole thing is pretty high variance. I really need to move to distributional RL to see if gets dampened any.
- Likewise I should test it using standard Q learning updates to see if that helps the variance any. I should test this at least once

Before I move to using RNNs, metaoptimization and the replay buffer, I really should first plug in distributional RL as it will make everything much easier.

It bothers me how poorly LN works and I am not sure why that is.

At any rate, my conclussion is that adding hidden layers does not really help performance, but I am really testing it for too low number of episodes to come to any firm conclussion.

Maybe on more complex games it would work better.

But I have the wrong idea in general. What I want is not to add layers, but to make it learn faster. And doing that is a metalearning problem. Or an architectural from another perspective.

I really should investigate what the problem with layer norm is. I'll try adding another layer and only using it in the first one.

Let me give a try and then I will take a break from this.