I am done. Adding more layers  just makes things worse. That fact that LN works worse is also interesting to me. The architecture of a NN is a prior on its learning.
If this makes it work poorly then there should be things to make it work better. I am just getting started with this.

The next two things in line would be to try distributional RL. I will try both quantile regression and the KL idea that I had originally.

I expect either of those to improve learning tremendously, but I am not sure how the two will compare against each other.

That will be my mission for the next few days and after that I'll get to work on the replay buffer. Reptile and RNNs will come after that.

I am done for the day. Let me go stretch out for a bit.