https://www.youtube.com/playlist?list=PLgKuh-lKre115020ug_gdmRroIfiQsPL4
I had a super early wake up today and have been reading and watched the Computational Assemblies on this playlist.

https://arxiv.org/abs/1804.10689
Decoupling Dynamics and Reward for Transfer Learning

I really like this paper for the reasons posted in this RL thread.

https://www.reddit.com/r/reinforcementlearning/comments/8gdtei/decoupling_dynamics_and_reward_for_transfer/

I'll make a little post on the PL sub in the monthly thread on what I've been up to.

Just like I have been for the past month, I seem to be going back and forth between studying and programming even though I've been leaning towards the later as time goes on. I think that this will be enough for today of studying. I'll try to get some programming done today.

I think I will just go with the QR idea and let my original one go. I do not want to waste time on useless experiments. Even if it works better by some chance, it won't be much better than QR.

I've been thinking about trading for the first time in a while.

Probably the main difficulty with trading is that is incredibly non trivial as a game - unlike poker for example, it cannot be reduced to an adversarial game.

You could put agents in a box and have them trade against each other, but all that would succeed is to teach them not to trade which is kind of ironic.

In order to properly place it in an adversarial context, it would be necessary to not just use price, but to trace the links all the way into the real world.

That is real tier three stuff or possible the beginner post-Singularity level. Yeah, sure Warren, trading on price does not make sense, though I am not sure that we have the same idea on what actually does.

Damn, I must have been wake for at least 4h at this point. Let me make a short post on the PL sub, and then I will take a break.