I spent a very good hour deep in focus on the QR paper. It seems there are quite some subtle things involved in its implementation, but it is quite easy.
Regardless, there are some things which stick out at me.

1) It never really occured to me before, but how am I supposed to use Q learning with recurrent nets.

It actually does not make sense unless I use MC updates or policy gradients.

As a matter of fact, I did think about this before, but somehow it completely slipped my mind that even with the next state, I cannot possibly leave the hidden state lying around.

Er, hopefully geri will come up with something good. If not, it does not matter, I'll just stick to MC updates for now.

2) QR-DQN is for deep Q learning which I am specifically not doing.

Using solely MC updates will simplify it significantly and not quite make it the same thing.

3) I should change reply_dq to reply_dmc to indicate that I am in fact doing Monte Carlo updates.

But I can in fact do Q learning updates if it is only in an online fashion.

I should test it, but I am lazy and do not feel like it. Especially today. TD learning does not really work better for me that MC even in the tabular case. I have no reason to think that things will be any different with function approximators.

4) Considering that PG style updates work in practice, there is no way that once I put in QR into MC, it will suddenly start working worse.

So I'll use QR with MC as my work horse.

5) The real issue with Q learning is the replay buffer. With online learning it makes no difference, but learning in an offline fashion will make using recurrent nets a lot more difficult unless I use MC updates.

Furthermore , in order to learn sequences, I want to be able to split the replay buffer neatly into chunks similarly to how I did the dataset for the char RNN.

So that leave MC updates as the only realistic option at the moment.

It might be possible to replace MC updates with something else (that is not Q learning) in the future.

6) That is about it. I guess the only thing that remains is to actually implement the thing. I understand it completely now whether that is how the cost function works or how to do action selection.

I'll start work on it later, maybe even tomorrow when I feel fresher. Right now let me go get some exercise.

I also feel like following up on some of the references from the paper I've linked to in the previous commit.