"4:40pm. https://arxiv.org/abs/1203.3497 Parametric Return Density Estimation for Reinforcement Learning
5pm. Got a reply.

https://www.reddit.com/r/reinforcementlearning/comments/8ghue5/d_how_are_the_q_values_calculated_with_quantile/dybra1z/?context=3&st=jgp84a80&sh=c46e04d5

So the quantiles are intended to be evenly split. I am now wondering how extreme values are handled.

6pm. Somehow I got hung up on this and reading papers on natural gradient for a long time. The day is over for me."

To be honest, I am a bit disappointed at the last few hours as I was very close to the answer as is, but the detail of quantile weights is not something that I could have just assumed.

At any rate, that about figures it. For the question in the reply to /u/wildML, I still haven't gotten an answer, but I am guessing the weight of the extreme value is probably 0. Or rather, I should just assume that they need to be evenly spaced and call it a day with that.

Without a doubt, tomorrow I will get to work on implementing this and trying it out. The replay buffer and recurrent nets are next in line after that.

It won't be long. Really, the difficult parts of this are like today: understanding the subtle details of the algorithms, deciding on the library design and writing efficient implementations.

Spiral is the one stop solution to the last problem and the rest is trivial once you understand it.