The video by Pollack was interesting, especially the part on how the teacher should be rewarded.
I will do what I am doing for now, but after I put in metaleraning and RNNs and everything else, I will seriously keep an eye out for how to crack hierarchical RL.

The biggest problem is that I do not understand the principle behind it. This is related to the fact that I do not exactly understand how to move from MC updates.

For today, let me just work on the selector for QR and its reverse. Baby steps and all that.