"1:50pm. Let me try the QR function on Mnist.
This will be a my last ditch effort to make sure whether it works.

2:10pm. Well, it works incredibly poorly. One thing I noticed that I was not aware of before is that LN with square error performs quite poorly. This sort of explains why I am having difficulties with it in poker.

But basically this a dead end. Maybe this thing would work with Q learning, but with MC updates like I am doing here it is awful. I did not expect this.

Plan B - let me use the discriminative KL loss. This was my original idea. Thankfully the new kernels will be just as convenient for that as they were for the previous thing.

While I am at it, I am going to rework the softmax cost a little so it recalculates its results. That should speed it up a bit.

I am absolutely sure the KL loss will work better than squared error.

2:25pm. Let me take a break here. I need to recharge my batteries."