I am going to have to rework the `mapi_d1_seq_broadcast` kernel in order to speed up the softmax cost.
That was not on my timetable, but I guess doing this stuff is inevitable if I want to attain perfection.

Well, let me sketch it out and then I will call it day.

I do not have the steam to do it completely today, but I'll make the foundations.

What I have to do is not too difficult so it might be doable for me to finish the KL agent in the next two days or maybe even tomorrow.