I am done for the day here. What I have here is sheer brilliance. I will replace the other `seq` with it, just like I did with `init`. And I will use this more generic function to implement softmax.
I'll just use atomics to add up the costs in the final layer. Also having this function on hand will make it easy for me implement the function that takes in a tensor of labels rather than a one hot encoding of them.

I am thinking that in order to streamline this, I could actually implement a new kind of tensor that does that sort of thing automatically.

That is just one of the ideas that I had. It is not a priority right now so forget it. There are various ways of doing this and I'd need to think of how I wanted to do it. It might be easier to just pass in a lambda somewhere.

https://arxiv.org/abs/1709.04054
Shifting Mean Activation Towards Zero with Bipolar Activation Functions

https://arxiv.org/abs/1707.07631
Deep Architectures for Neural Machine Translation

https://arxiv.org/abs/1507.00210
Natural Neural Networks

https://www.reddit.com/r/MachineLearning/comments/8gpp0k/d_where_is_evidence_that_batch_normalization/?st=jgs40s7i&sh=f669e918

There is an interesting discussion going on in this thread, so I might as well do some reading.

I'll continue this tomorrow.