I've finished rediffing the cache. I got rid of learning6, learning7 and learning8 as the 3 tests are covered by learning9 in their entirety and I do not feel like messing with them now.
With tihs, I can make the change to forbid the updates over the layer types.

Agh, do I want to forbid these kinds of updates or do I want to go what I had previously? Do I want consistency or usefulness?

Well, I can restrict it now and revert the decision at a later point if it turns out to be wrong. But if I allow it and it turns out to be wrong, I won't be able to take it back. So let me move towards restricting it.