Removed the support for layout types in the `ModuleWith` construct.
Originally this feature was intended to be a convenience, but now after half a year of practical experience with them I've realized that layout types are simply never used that way so they are better off being restricted. It will make the construct more consistent with the rest of the language.

It would be easy to add it should the need arise.