I've gotten rid of the call to zero and now the two methods are up to par. Strange that speed did not improve, but at least this will make memory usage lower.
Ok...phew. I am taking a short breather at the moment.

I hate bugs like these as they waste so much of my time. There is actually a memory race in run_parallel since memory transfers are async, but it won't be a problem in practice.

I really hate Cuda. The separation of GPU and the main memory is source of so much complexity.

A lot of inefficiency could vanish if the GPU were more flexible in its capabilities and more closely integrated with the CPU.

Fuuuu....ok, what is next?

I think it would be to start work on the selector for KL cost. Just like before, I'll do this one by one.